+++ src/app/(dashboard)/components/IngredientList.tsx
@@ -1,0 +1,18 @@
+import { IngredientAmount } from '@/types/Ingredient'
+import { Avatar, ListItem, ListItemAvatar, ListItemText } from '@mui/material'
+import EggIcon from '@mui/icons-material/Egg'
+
+export const IngredientList = ({ ingredients }: { ingredients: IngredientAmount[] }) =>
+    ingredients.map(({ ingredient, amount, unit }, id) => (
+        <ListItem key={id}>
+            <ListItemAvatar>
+                <Avatar>
+                    <EggIcon />
+                </Avatar>
+            </ListItemAvatar>
+            <ListItemText
+                primary={ingredient.name}
+                secondary={`Ilość: ${amount} ${unit}`}
+            />
+        </ListItem>
+    ))
