+++ src/app/(dashboard)/layout.tsx
@@ -2,9 +2,10 @@ import '@/app/globals.css'
 import type { Metadata } from 'next'
 import { AppProvider } from '@toolpad/core/AppProvider'
 import { ReactNode } from 'react'
-import { MenuBar } from '@/app/(dashboard)/MenuBar'
+import { MenuBar } from '@/components/MenuBar'
 import { appName } from '@/constants/general'
 import { NotificationsProvider } from '@toolpad/core'
+import { Box } from '@mui/material'
 
 export const metadata: Metadata = {
     title: `Panel - ${appName}`,
@@ -24,7 +25,7 @@ const DashboardLayout = ({
                 },
             }}
         >
-            <div
+            <Box
                 style={{
                     maxHeight: '100vh',
                     minHeight: '100vh',
@@ -32,7 +33,7 @@ const DashboardLayout = ({
             >
                 <MenuBar />
                 {children}
-            </div>
+            </Box>
         </NotificationsProvider>
     </AppProvider>
 )
