+++ src/constants/general.ts
@@ -1,1 +1,2 @@
 export const appName = 'nieDOBRE.com'
+export const autoHideDuration = 6000
